+++ lib/operations/retry.ts
@@ -4,8 +4,9 @@ import { wrap } from './wrap';
 
 interface Options {
   count?: number;
-  onTryError?: AnyFunction;
-  onFinalError?: AnyFunction;
+  onTryError?: (error: Error) => any;
+  onFinalError?: (error: Error) => any;
+  shouldRetry?: (error: Error) => boolean | Promise<boolean>;
 }
 
 export function rawRetryWithOptions<F extends AnyFunction>(
@@ -15,43 +16,62 @@ export function rawRetryWithOptions<F extends AnyFunction>(
   const {
     count = 3,
     onTryError = () => undefined,
-    onFinalError = () => undefined
+    onFinalError = () => undefined,
+    shouldRetry = () => true
   } = options;
   const wrappedOnTryError = wrap(onTryError);
   const wrappedOnFinalError = wrap(onFinalError);
+  const wrappedshouldRetry = wrap(shouldRetry);
   const wrappedFunc = wrap(func);
   return doRawRetryWithOptions(
     {
       count,
       onTryError: wrappedOnTryError,
-      onFinalError: wrappedOnFinalError
+      onFinalError: wrappedOnFinalError,
+      shouldRetry: wrappedshouldRetry
     },
     wrappedFunc
   );
 }
 
+interface ResolvedOptions {
+  count: number;
+  onTryError: (error: Error) => Promise<any>;
+  onFinalError: (error: Error) => Promise<any>;
+  shouldRetry: (error: Error) => Promise<boolean>;
+}
+
 function doRawRetryWithOptions<F extends AnyFunction>(
-  options: Required<Options>,
+  options: Required<ResolvedOptions>,
   func: F
 ): PromisifiedFunction<F> {
-  const { count, onTryError, onFinalError } = options;
+  const { count, onTryError, onFinalError, shouldRetry } = options;
   return (...args: Parameters<F>) =>
     func(...args).catch((error: Error) => {
       const nextCount = count - 1;
-      if (nextCount < 0) {
-        return onFinalError(error).then(() => Promise.reject(error));
-      }
-      return onTryError(error).then(() =>
-        doRawRetryWithOptions(
-          {
-            count: nextCount,
-            onTryError,
-            onFinalError
-          },
-          func
-        )(...args)
-      );
+      return canRetry(error, nextCount).then(retryPossible => {
+        if (!retryPossible) {
+          return onFinalError(error).then(() => Promise.reject(error));
+        }
+        return onTryError(error).then(() =>
+          doRawRetryWithOptions(
+            {
+              count: nextCount,
+              onTryError,
+              onFinalError,
+              shouldRetry
+            },
+            func
+          )(...args)
+        );
+      });
     });
+
+  function canRetry(error: Error, nextCount: number): Promise<boolean> {
+    return shouldRetry(error).then((condition: boolean) => {
+      return condition && nextCount >= 0;
+    });
+  }
 }
 
 const curriedRetryWithOptions = autoCurry(rawRetryWithOptions);